Refactored ghofrani e macias
