+++ openquake/hazardlib/gsim/nshmp_2014.py
@@ -22,36 +22,6 @@ Module exports :class:`AtkinsonMacias2009NSHMP2014` and :class:`NSHMP2014`
 import numpy as np
 from openquake.hazardlib import const
 from openquake.hazardlib.gsim import base
-from openquake.hazardlib.gsim.atkinson_macias_2009 import AtkinsonMacias2009
-from openquake.hazardlib.gsim.can15.sinter import SInterCan15Mid
-
-
-class AtkinsonMacias2009NSHMP2014(AtkinsonMacias2009):
-    """
-    Implements an adjusted version of the Atkinson and Macias (2009) GMPE.
-    The motion is scaled B/C conditions following the approach described in
-    Atkinson and Adams (2013) and implemented in
-    :mod:`openquake.hazardlib.gsim.can15.sinter`.
-    """
-
-    #: Shear-wave velocity for reference soil conditions in [m s-1]
-    DEFINED_FOR_REFERENCE_VELOCITY = 760.
-
-    #: GMPE not tested against independent implementation so raise
-    #: not verified warning
-    non_verified = True
-
-    def get_mean_and_stddevs(self, sctx, rctx, dctx, imt, stddev_types):
-        """
-        See :meth:`superclass method
-        <.base.GroundShakingIntensityModel.get_mean_and_stddevs>`
-        for spec of input and result values.
-        """
-        # Get original mean and standard deviations
-        mean, stddevs = super().get_mean_and_stddevs(
-            sctx, rctx, dctx, imt, stddev_types)
-        mean += np.log(SInterCan15Mid.COEFFS_SITE[imt]['mf'])
-        return mean, stddevs
 
 
 def nga_west2_epistemic_adjustment(magnitude, distance):
